css/css-contain/container-queries/container-units-svglength.html

<!DOCTYPE html>
<title>CSS Container Queries Test: container-relative units in SVGLength</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-lengths">
<link rel="help" href="https://svgwg.org/svg2-draft/types.html#InterfaceSVGLength">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #container {
    container-type: size;
    width: 200px;
    height: 150px;
  }
</style>
<div id=container>
  <svg>
    <rect id="rect1" width="10cqw" height="10cqh"/>
    <rect id="rect2" width="10cqi" height="10cqb"/>
    <rect id="rect3" width="10cqmin" height="10cqmax"/>
    <rect id="rect4" width="calc(10cqmin + 10cqmax)" height="calc(10cqw + 3px)"/>
  </svg>
</div>
<script>
  setup(() => {
    assert_implements_container_queries();
    container.offsetTop;
  });

  test(() => {
    assert_equals(rect1.width.baseVal.value, 20);
    assert_equals(rect1.height.baseVal.value, 15);
  }, 'cqw,cqh can be resolved');

  test(() => {
    assert_equals(rect2.width.baseVal.value, 20);
    assert_equals(rect2.height.baseVal.value, 15);
  }, 'cqi,cqb can be resolved');

  test(() => {
    assert_equals(rect3.width.baseVal.value, 15);
    assert_equals(rect3.height.baseVal.value, 20);
  }, 'cqmin,cqmax can be resolved');

  test(() => {
    assert_equals(rect4.width.baseVal.value, 35);
    assert_equals(rect4.height.baseVal.value, 23);
  }, 'calc() with container-relative units can be resolved');
</script>
